feat: finished implementing auth module
